+++ actix-cors/CHANGES.md
@@ -1,7 +1,10 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+* Make `Cors` middleware generic over body type [#195]
 * Minimum supported Rust version (MSRV) is now 1.51.
+
+[#195]: https://github.com/actix/actix-extras/pull/195
 
 
 ## 0.6.0-beta.2 - 2021-06-27